ai-agent/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.1
python-telegram-bot>=21.0
httpx>=0.27.0
jinja2>=3.1.4
weasyprint>=63.0
aiosqlite>=0.20.0
